refactor(config_routes): extract learning.py + autonomous.py + social.py — split complete (8/8)
Schritt 8 (final) im Split-Plan
(`docs/superpowers/plans/2026-04-29-config-routes-split.md`).

What moved
- `_register_learning_routes()` (~440 LOC) -> `cognithor.channels.config_routes.learning`. Active Learning, Curiosity-Engine, Confidence-Manager, Reflection-Engine.
- `_register_ingest_routes()` (~196 LOC) -> selbe Datei. Knowledge-Ingestion (File/URL/YouTube), Job-Queue.
- `_register_autonomous_routes()` (~20 LOC) -> `cognithor.channels.config_routes.autonomous`. Autonomous-Task-Orchestration.
- `_register_feedback_routes()` (~137 LOC) -> selbe Datei. Thumbs-up/down + Chat-Tree-Vote.
- `_register_social_routes()` (~328 LOC) -> `cognithor.channels.config_routes.social`. LeadService / Reddit Lead Hunter.

`_factory.py` ist mit dieser PR fertig: **119 Zeilen** (von ursprunglich 6 628). Reine Orchestrator-Funktion `create_config_routes()` plus `_get_hub`-Closure. 24 `_register_*_routes()`-Helfer leben jetzt verteilt auf 14 Sub-Module.

`_factory.py` Cleanup
- Docstring umgeschrieben (Refactor abgeschlossen, kein "wird aufgeteilt" mehr).
- Top-Level `Request`/`HTTPException` Try-Imports entfernt — Factory hat keine eigenen Endpoints mehr.
- `import contextlib` entfernt (war fuer learning).
- `from cognithor.utils.logging import get_logger` und `log = get_logger(__name__)` entfernt — Factory loggt nicht mehr.

Public API
`from cognithor.channels.config_routes import create_config_routes` ist seit PR 1 unveraendert; alle 11 Konsumenten (2x `__main__.py`, 2x `webui.py`, 9 Test-Files) blieben ueber den gesamten Refactor unangefasst.

Cumulative Result PRs 1-8
- _factory.py: 6 628 -> 119 Zeilen (-98 %).
- Sub-Module: 14 (autonomous, config, evolution, governance, infrastructure, learning, monitoring, security, session, skills, social, system, ui, workflows).
- Endpoints: 303/303 unveraendert (Drift-Test in jedem PR gruen).
- Konsumenten-Tests: 396/396 in jedem PR gruen.

Verification
- `pytest test_route_inventory_unchanged` -> passed.
- Konsumenten-Sweep ueber 11 Test-Files -> 396 passed.
- `ruff check` -> All checks passed.
- `ruff format --check` -> 16 files unchanged.
